Move dump net buffers to udp module
