+++ app/gimlet/rev-b.toml
@@ -41,7 +41,7 @@ name = "task-net"
 stacksize = 6040
 priority = 5
 features = ["mgmt", "h753", "gimlet", "vlan", "vpd-mac"]
-max-sizes = {flash = 131072, ram = 32768, sram1 = 16384}
+max-sizes = {flash = 131072, ram = 65536, sram1 = 16384}
 sections = {eth_bulk = "sram1"}
 uses = ["eth", "eth_dma", "tim16"]
 start = true
@@ -309,12 +309,14 @@ task-slots = ["sys"]
 
 [tasks.dump_agent]
 name = "task-dump-agent"
-priority = 5
-max-sizes = {flash = 8192, ram = 2048 }
+priority = 6
+max-sizes = {flash = 16384, ram = 8192 }
 start = true
-task-slots = ["sprot", "jefe"]
-stacksize = 1200
+task-slots = ["sprot", "jefe", "net"]
+stacksize = 2400
 extern-regions = ["sram2", "sram3", "sram4"]
+notifications = ["socket"]
+features = ["net", "vlan"]
 
 [tasks.sbrmi]
 name = "drv-sbrmi"
@@ -1283,6 +1285,13 @@ owner = {name = "control_plane_agent", notification = "socket"}
 port = 11111 # TODO do we have a documented port for MGS traffic?
 tx = { packets = 3, bytes = 2048 }
 rx = { packets = 3, bytes = 2048 }
+
+[config.net.sockets.dump_agent]
+kind = "udp"
+owner = {name = "dump_agent", notification = "socket"}
+port = 11113 # TODO do we have a documented port for dump_agent traffic?
+tx = { packets = 3, bytes = 1024 }
+rx = { packets = 3, bytes = 1024 }
 
 [config.sprot]
 # ROT_IRQ (af=0 for GPIO, af=15 when EXTI is implemneted)
